長片記憶體:音檔改用 float32 讀,而且整條管線只讀一次
soundfile 預設是 float64,而這條管線會把整支影片的音訊一次讀進記憶體。
更糟的是同一個檔被讀了兩遍——音樂偵測與微剪各自 sf.read 一次。
再加上 _window_db 裡的 trimmed.astype(np.float64) ** 2 會複製出兩份跟音檔
一樣大的暫存陣列,實際峰值遠比想像中高。

單聲道 48kHz 換算:
    34 分鐘  float64 783MB/次  ->  float32 392MB/次
    90 分鐘  float64 2.1GB/次  ->  float32 1.0GB/次
而且 Whisper 剛跑完,那個時間點記憶體本來就吃緊。

三處一起改:讀取指定 float32、pipeline 讀一次分給兩個偵測、
_window_db 改用 np.mean(..., dtype=np.float64) 讓 numpy 直接用 float64 累加
(精度一樣,但少掉那兩份整份複製)。gate_speed_audio 也一併改。
實測 10 分鐘的音檔,峰值從 462MB 降到 231MB。

---

重點是驗了「我沒有在優化的那一面」。記憶體降多少是我在優化的數字,
真正該問的是「精度變低會不會改到剪輯結果」——如果某段音樂因此掉到門檻
下面,那段音樂就會被剪掉,而這種事在報告上看不出來,只有播出去才發現。

拿一段合成的「像真實教學片」的音訊(講話、停頓、示範音樂、咳嗽、底噪),
寫成 16-bit PCM 再分別用 float64 / float32 讀回來,逐項比對:
    逐窗 dB 值   12071 窗 + 30178 窗,最大差 7.6e-08 dB(門檻是「分貝」等級)
    音樂/有聲區間  47 段 vs 47 段,完全一致
    微剪安靜區間   54 段 vs 54 段,完全一致
    gate 的讀寫來回  寫出的檔案位元組完全相同(subtype 仍是 PCM_16)

這個比對留成測試 test_float32_gives_identical_detection,以後有人再動精度
會當場紅掉。另外加 test_probe_reads_audio_once_as_float32 守住「只讀一次」
與「不要又把整份轉回 float64」。

+++ modules/audio_clean.py
@@ -338,7 +338,10 @@ def gate_speed_audio(in_wav: str, out_wav: str, segments, fps: float) -> str:
     原理:在 Premiere 裡快轉播放時,若那段本來就無聲,加速後仍是無聲,
     就不會有加速造成的尖聲(花栗鼠音)。回傳處理後的 WAV 路徑。"""
     import soundfile as sf
-    audio, sr = sf.read(in_wav)
+    # float32 而不是預設的 float64:整支影片的音訊會一次讀進記憶體,
+    # 34 分鐘的片這一項就差 390MB。抹靜音只是把樣本設成 0,不做運算,
+    # 而寫出去仍然是 16-bit PCM——已實測兩者寫出的檔案位元組完全相同。
+    audio, sr = sf.read(in_wav, dtype="float32")
     n = len(audio)
     muted = 0
     for s in segments:
